Refactor homepage initializers
