+++ assets/js/utils/data-attributes.ts
@@ -1,0 +1,15 @@
+export const DATASET_KEYS = {
+  quickstartSlug: 'quickstartSlug',
+};
+
+export const DATA_SELECTORS = {
+  quickstart: '[data-quickstart-slug]',
+  topNav: '[data-top-nav]',
+  previewReel: '[data-preview-reel]',
+  previewCard: '[data-preview-card]',
+  previewDot: '[data-preview-dot]',
+  previewStatus: '[data-preview-status]',
+  previewPrev: '[data-preview-prev]',
+  previewNext: '[data-preview-next]',
+  previewToggle: '[data-preview-toggle]',
+};
